# swagger-document.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/schedule/branchs/paginate": {
      "get": {
        "operationId": "BranchsController_paginate",
        "parameters": [
          {
            "name": "branch_name",
            "required": false,
            "in": "query",
            "description": "Name of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cnpj",
            "required": false,
            "in": "query",
            "description": "CNPJ of the branch.",
            "example": "12345678000200",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "street",
            "required": false,
            "in": "query",
            "description": "Street of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cep",
            "required": false,
            "in": "query",
            "description": "CEP of the branch.",
            "example": "36150000",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "City of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "district",
            "required": false,
            "in": "query",
            "description": "District of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "local_number",
            "required": false,
            "in": "query",
            "description": "Number residence of the branch.",
            "example": "230B",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch_phone",
            "required": false,
            "in": "query",
            "description": "Number phone of the branch.",
            "example": "32227460",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "complements",
            "required": false,
            "in": "query",
            "description": "Information additional of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "opening_hours",
            "required": false,
            "in": "query",
            "description": "Start of branch office hours",
            "example": "08:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "closing_hours",
            "required": false,
            "in": "query",
            "description": "End of branch office hours",
            "example": "08:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Owner of the branch.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "Date create of the branch.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "Date update of the branch.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "Date delete of the branch.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite data of the paginate branchs.",
            "example": 100,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page of the paginate branchs.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "branchs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/schedule/branchs/{id}": {
      "get": {
        "operationId": "BranchsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        },
        "tags": [
          "branchs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "BranchsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        },
        "tags": [
          "branchs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "BranchsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "branchs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/schedule/branchs": {
      "get": {
        "operationId": "BranchsController_findAll",
        "parameters": [
          {
            "name": "branch_name",
            "required": false,
            "in": "query",
            "description": "Name of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cnpj",
            "required": false,
            "in": "query",
            "description": "CNPJ of the branch.",
            "example": "12345678000200",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "street",
            "required": false,
            "in": "query",
            "description": "Street of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cep",
            "required": false,
            "in": "query",
            "description": "CEP of the branch.",
            "example": "36150000",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "City of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "district",
            "required": false,
            "in": "query",
            "description": "District of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "local_number",
            "required": false,
            "in": "query",
            "description": "Number residence of the branch.",
            "example": "230B",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch_phone",
            "required": false,
            "in": "query",
            "description": "Number phone of the branch.",
            "example": "32227460",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "complements",
            "required": false,
            "in": "query",
            "description": "Information additional of the branch.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "opening_hours",
            "required": false,
            "in": "query",
            "description": "Start of branch office hours",
            "example": "08:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "closing_hours",
            "required": false,
            "in": "query",
            "description": "End of branch office hours",
            "example": "08:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Owner of the branch.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "Date create of the branch.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "Date update of the branch.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "Date delete of the branch.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite data of the paginate branchs.",
            "example": 100,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page of the paginate branchs.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "branchs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "BranchsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        },
        "tags": [
          "branchs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Schedule API example",
    "description": "The schedule API description",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          },
          "branchs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Branch"
            }
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "user_name",
          "password",
          "user_email",
          "phone_number",
          "branchs",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Client": {
        "type": "object",
        "properties": {
          "client_name": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "birth_date": {
            "type": "string"
          },
          "client_phone": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "branch_id": {
            "type": "number"
          },
          "branch": {
            "$ref": "#/components/schemas/Branch"
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "client_name",
          "first_name",
          "birth_date",
          "client_phone",
          "is_active",
          "branch_id",
          "branch",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Service": {
        "type": "object",
        "properties": {
          "service_name": {
            "type": "string"
          },
          "service_value": {
            "type": "number"
          },
          "expected_time": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "user_id": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "branchs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Branch"
            }
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "service_name",
          "service_value",
          "expected_time",
          "is_active",
          "user_id",
          "user",
          "branchs",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Branch": {
        "type": "object",
        "properties": {
          "branch_name": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "cep": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "user_id": {
            "type": "number"
          },
          "district": {
            "type": "string"
          },
          "local_number": {
            "type": "string"
          },
          "branch_phone": {
            "type": "string"
          },
          "complements": {
            "type": "string"
          },
          "opening_hours": {
            "type": "string"
          },
          "closing_hours": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Client"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "branch_name",
          "cnpj",
          "street",
          "cep",
          "city",
          "user_id",
          "district",
          "local_number",
          "branch_phone",
          "complements",
          "opening_hours",
          "closing_hours",
          "user",
          "clients",
          "services",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "branch_name": {
            "type": "string",
            "description": "Name of the branch."
          },
          "cnpj": {
            "type": "string",
            "description": "CNPJ of the branch.",
            "example": "12345678000200"
          },
          "street": {
            "type": "string",
            "description": "Street of the branch."
          },
          "cep": {
            "type": "string",
            "description": "CEP of the branch.",
            "example": "36150000"
          },
          "city": {
            "type": "string",
            "description": "City of the branch."
          },
          "district": {
            "type": "string",
            "description": "District of the branch."
          },
          "local_number": {
            "type": "string",
            "description": "Number residence of the branch.",
            "example": "230B"
          },
          "branch_phone": {
            "type": "string",
            "description": "Number phone of the branch.",
            "example": "32227460"
          },
          "complements": {
            "type": "string",
            "description": "Information additional of the branch."
          },
          "opening_hours": {
            "type": "string",
            "description": "Start of branch office hours",
            "example": "08:00"
          },
          "closing_hours": {
            "type": "string",
            "description": "End of branch office hours",
            "example": "08:00"
          },
          "user_id": {
            "type": "number",
            "description": "Affiliate user id (owner).",
            "example": 1
          },
          "services": {
            "description": "Services of the branch.",
            "example": "[{id: 1}]",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "branch_name",
          "cnpj",
          "street",
          "cep",
          "city",
          "district",
          "local_number",
          "opening_hours",
          "closing_hours",
          "user_id"
        ]
      },
      "UpdateBranchDto": {
        "type": "object",
        "properties": {
          "branch_name": {
            "type": "string",
            "description": "Name of the branch."
          },
          "cnpj": {
            "type": "string",
            "description": "CNPJ of the branch.",
            "example": "12345678000200"
          },
          "street": {
            "type": "string",
            "description": "Street of the branch."
          },
          "cep": {
            "type": "string",
            "description": "CEP of the branch.",
            "example": "36150000"
          },
          "city": {
            "type": "string",
            "description": "City of the branch."
          },
          "district": {
            "type": "string",
            "description": "District of the branch."
          },
          "local_number": {
            "type": "string",
            "description": "Number residence of the branch.",
            "example": "230B"
          },
          "branch_phone": {
            "type": "string",
            "description": "Number phone of the branch.",
            "example": "32227460"
          },
          "complements": {
            "type": "string",
            "description": "Information additional of the branch."
          },
          "opening_hours": {
            "type": "string",
            "description": "Start of branch office hours",
            "example": "08:00"
          },
          "closing_hours": {
            "type": "string",
            "description": "End of branch office hours",
            "example": "08:00"
          },
          "user_id": {
            "type": "number",
            "description": "Affiliate user id (owner).",
            "example": 1
          },
          "services": {
            "description": "Services of the branch.",
            "example": "[{id: 1}]",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  }
}